Use a flag to show which view is "active"
There can always be only one selected item, either a history clip or a favourite.
To keep track which of the two lists is the "active" one that colours the currently selected item, the MainWindow manages the flag fHistoryActiveFlag.
view->IsFocus() can't be used, because when a menu is invoked both lists lose focus.