Allow to remove the only clip left (= current clipboard)
Clipdinger used to silently refuse to remove the last clip left
in the history (which, being top in the list, is what's currently
in the clipboard).

We allow to remove the top clip in all other situations: then the
clip below will be the new top of the list and becomes the new
clibboard contents.

Seems logical to clear the clipboard when the last remaining, top
clip is deleted.

"Clear clip history" will still keep the top clip, because removing
that would also wipe the current clipboard contents not just the
history.

Fixes #30