Fix inserted items not moving to top when filtering
Normally, if you insert a clip from further down the history list,
it will be moved up to the top of the list.

When filtering was introduced, this broke when inserting a clip from
the filtered list, because that is a separate temporary list and any
changes there (position or updated colours) will be forgotten when
the previously backed-up fHistory list gets restored.

We now treat the inserted clip like a newly copied clip by not stopping
the watch of the clipboard when putting the clip into it. That will
trigger B_CLIPBOARD_CHANGED and consequently remove the old clip via
_MakeItemUnique() and move it up with MoveClipToTop().