Fix stopping the input device
A recent change in Haiku exposed a bug in the InputDevice part of
Clipdinger. It affected updating Clipdinger, because removing the
InputDevice in order to replace it with the new version, pretty
much froze the system, see: https://dev.haiku-os.org/ticket/19100

The InputDevice was never stopped as it was stuck in a loop reading
from a port. We now delete the port first, before waiting on the
listener thread to stop.

Thanks to madmax for pointing out the problem and suggesting this
fix.

Fixes #32